Implement Database Library for Persistent Local Storage
A lot of the times, we need to extract information from Popmundo.
A lot of the times, the information will not change.
This library is for keeping track of that information so we don't make
unnecessary request to Popmundo for stuff we already know.

TODO: Also add when the info was added for expiring information.